add pagination for calls sections (#460)
