+++ app/utils/cjs/lodash.js
@@ -4,4 +4,5 @@ import orderBy from 'lodash/orderBy.js'
 import pick from 'lodash/pick.js'
 import shuffle from 'lodash/shuffle.js'
 import sortBy from 'lodash/sortBy.js'
-export {isEqual, omit, orderBy, pick, shuffle, sortBy}
+import groupBy from 'lodash/groupBy.js'
+export {isEqual, omit, orderBy, pick, shuffle, sortBy, groupBy}
